AppletContainer: Fix showing busy indicator directly after applet is created
There is a chance for the applet to mark itself as busy before the AppletContainer
starts listening to Applet::busyChanged. In this case, the busy indicator doesn't show up
until busy is toggled off and on again.

To fix this, run the code that is run on busyChanged also when initializing.